backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
python-dotenv==1.0.1

azure-storage-blob==12.23.1
azure-ai-formrecognizer==3.3.3
azure-cosmos==4.7.0

openai==1.52.2
httpx<0.28

pypdf==5.1.0
python-docx==1.1.2
